[update] AppStorage for setting persistance

+++ NASPlayer/Views/Album/AlbumSongListView.swift
@@ -41,7 +41,7 @@ struct AlbumDiscView: View {
     var tracks: [SongVM]?
     
     var safeWidth: CGFloat {
-        let logicalWidth: CGFloat = CGFloat(Device.current.logicalResolution?.width ?? 0.0)
+        let logicalWidth: CGFloat = Device.current.logicalResolution?.width ?? 0.0
         return logicalWidth - 120
     }
