+++ NASPlayer/Views/Browse/Albums/AlbumCoverView.swift
@@ -18,7 +18,7 @@ struct AlbumCoverView: View {
     }
     
     var cardSize: CGFloat {
-        let logicalWidth: CGFloat = CGFloat(Device.current.logicalResolution?.width ?? 0.0)
+        let logicalWidth: CGFloat = Device.current.logicalResolution?.width ?? 0.0
         return (logicalWidth - 56) / 2
     }
