+++ NASPlayer/Views/Playlist/PlaylistEntryView.swift
@@ -18,7 +18,7 @@ struct PlaylistEntryView: View {
     }
     
     var safeWidth: CGFloat {
-        let logicalWidth: CGFloat = CGFloat(Device.current.logicalResolution?.width ?? 0.0)
+        let logicalWidth: CGFloat = Device.current.logicalResolution?.width ?? 0.0
         return logicalWidth - 120
     }
