+++ NASPlayer/Views/Playlist/PlaylistSongListView.swift
@@ -37,7 +37,7 @@ struct PlaylistSongView: View {
     }
     
     var safeWidth: CGFloat {
-        let logicalWidth: CGFloat = CGFloat(Device.current.logicalResolution?.width ?? 0.0)
+        let logicalWidth: CGFloat = Device.current.logicalResolution?.width ?? 0.0
         return logicalWidth - 120
     }
